+++ apps/api/src/app/inbox/usecases/session/session.command.ts
@@ -1,5 +1,8 @@
-import { IsDefined, IsOptional, IsString } from 'class-validator';
+import { IsDefined, IsOptional, IsString, ValidateNested } from 'class-validator';
 import { BaseCommand } from '@novu/application-generic';
+import { Type } from 'class-transformer';
+
+import { SubscriberDto } from '../../dtos/subscriber-session-request.dto';
 
 export class SessionCommand extends BaseCommand {
   @IsDefined()
@@ -11,8 +14,9 @@ export class SessionCommand extends BaseCommand {
   readonly subscriberHash?: string;
 
   @IsDefined()
-  @IsString()
-  readonly subscriberId: string;
+  @ValidateNested()
+  @Type(() => SubscriberDto)
+  readonly subscriber: SubscriberDto;
 
   @IsOptional()
   @IsString()