+++ apps/api/src/app/inbox/usecases/session/session.spec.ts
@@ -66,7 +66,9 @@ describe('Session', () => {
   it('should throw an error if the environment is not found', async () => {
     const command: SessionCommand = {
       applicationIdentifier: 'invalid-app-id',
-      subscriberId: 'subscriber-id',
+      subscriber: {
+        subscriberId: 'subscriber-id',
+      },
     };
 
     environmentRepository.findEnvironmentByIdentifier.resolves(null);
@@ -82,7 +84,9 @@ describe('Session', () => {
   it('should throw an error if the in-app integration is not found', async () => {
     const command: SessionCommand = {
       applicationIdentifier: 'app-id',
-      subscriberId: 'subscriber-id',
+      subscriber: {
+        subscriberId: 'subscriber-id',
+      },
     };
 
     environmentRepository.findEnvironmentByIdentifier.resolves({
@@ -103,7 +107,9 @@ describe('Session', () => {
   it('should validate HMAC encryption and return the session response', async () => {
     const command: SessionCommand = {
       applicationIdentifier: 'app-id',
-      subscriberId: 'subscriber-id',
+      subscriber: {
+        subscriberId: 'subscriber-id',
+      },
       subscriberHash: 'hash',
     };
     const subscriber = { _id: 'subscriber-id' };
@@ -132,7 +138,9 @@ describe('Session', () => {
   it('should return correct removeNovuBranding value when is set on the integration', async () => {
     const command: SessionCommand = {
       applicationIdentifier: 'app-id',
-      subscriberId: 'subscriber-id',
+      subscriber: {
+        subscriberId: 'subscriber-id',
+      },
       subscriberHash: 'hash',
     };
     const subscriber = { _id: 'subscriber-id' };
@@ -177,7 +185,9 @@ describe('Session', () => {
   it('should create a subscriber and return the session response', async () => {
     const command: SessionCommand = {
       applicationIdentifier: 'app-id',
-      subscriberId: 'subscriber-id',
+      subscriber: {
+        subscriberId: 'subscriber-id',
+      },
       subscriberHash: 'hash',
       origin: 'origin',
     };
@@ -211,7 +221,7 @@ describe('Session', () => {
   it('should return the correct maxSnoozeDurationHours value for different service levels', async () => {
     const command: SessionCommand = {
       applicationIdentifier: 'app-id',
-      subscriberId: 'subscriber-id',
+      subscriber: { subscriberId: 'subscriber-id' },
       subscriberHash: 'hash',
     };