+++ apps/api/src/app/inbox/utils/encryption.ts
@@ -1,5 +1,5 @@
 import { BadRequestException } from '@nestjs/common';
-import { createHash, decryptApiKey } from '@novu/application-generic';
+import { isHmacValid } from '../../shared/helpers/is-valid-hmac';
 
 export function validateHmacEncryption({
   apiKey,
@@ -10,9 +10,7 @@ export function validateHmacEncryption({
   subscriberId: string;
   subscriberHash?: string;
 }) {
-  const key = decryptApiKey(apiKey);
-  const hmacHash = createHash(key, subscriberId);
-  if (hmacHash !== subscriberHash) {
+  if (!isHmacValid(apiKey, subscriberId, subscriberHash)) {
     throw new BadRequestException('Please provide a valid HMAC hash');
   }
 }